elk/python_apps_with_db/client/requirements.txt
requests==2.32.5
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-exporter-otlp==1.40.0
